+++ src/or/include.am
@@ -103,6 +103,7 @@ LIBTOR_A_SOURCES = \
 	src/or/scheduler.c				\
 	src/or/scheduler_kist.c				\
 	src/or/scheduler_vanilla.c			\
+	src/or/shared_random_common.c			\
 	src/or/statefile.c				\
 	src/or/status.c					\
 	src/or/torcert.c				\
@@ -260,6 +261,7 @@ ORHEADERS = \
 	src/or/routerset.h				\
 	src/or/routerparse.h				\
 	src/or/scheduler.h				\
+	src/or/shared_random_common.h			\
 	src/or/statefile.h				\
 	src/or/status.h					\
 	src/or/torcert.h				\
