+++ src/or/routerparse.c
@@ -74,6 +74,7 @@
 #include "entrynodes.h"
 #include "torcert.h"
 #include "sandbox.h"
+#include "shared_random_common.h"
 #include "dirauth/shared_random.h"
 
 #undef log
